+++ src/meltano/core/setting_definition.py
@@ -1,10 +1,12 @@
 import json
 from typing import List
 
-from .utils import truthy, flatten
+from .utils import truthy, flatten, nest_object
 from .behavior.canonical import Canonical
 from .behavior import NameEq
 from .error import Error
+
+VALUE_PROCESSORS = {"nest_object": nest_object}
 
 
 class SettingMissingError(Error):
@@ -33,6 +35,7 @@ class SettingDefinition(NameEq, Canonical):
         protected: bool = None,
         env_specific: bool = None,
         custom: bool = False,
+        value_processor=None,
         **attrs,
     ):
         super().__init__(
@@ -52,6 +55,7 @@ class SettingDefinition(NameEq, Canonical):
             placeholder=placeholder,
             protected=protected,
             env_specific=env_specific,
+            value_processor=value_processor,
             _custom=custom,
             **attrs,
         )
@@ -125,5 +129,11 @@ class SettingDefinition(NameEq, Canonical):
                 value = json.loads(value)
                 if not isinstance(value, list):
                     raise ValueError(f"JSON value '{value}' is not an array")
+
+        processor = self.value_processor
+        if value is not None and processor:
+            if isinstance(processor, str):
+                processor = VALUE_PROCESSORS[processor]
+            value = processor(value)
 
         return value